# tac: correct behavior of -b option