--- a/src/uu/tac/src/tac.rs
+++ b/src/uu/tac/src/tac.rs
@@ -11,7 +11,7 @@
 extern crate uucore;
 
 use clap::{crate_version, App, Arg};
-use std::io::{stdin, stdout, BufReader, Read, Stdout, Write};
+use std::io::{stdin, stdout, BufReader, Read, Write};
 use std::{fs::File, path::Path};
 use uucore::InvalidEncodingHandling;
 
@@ -84,11 +84,48 @@ pub fn uu_app() -> App<'static, 'static> {
         .arg(Arg::with_name(options::FILE).hidden(true).multiple(true))
 }
 
-fn tac(filenames: Vec<String>, before: bool, _: bool, separator: &str) -> i32 {
-    let mut exit_code = 0;
+fn buffer_tac(data: &[u8], before: bool, separator: &str) -> std::io::Result<()> {
     let mut out = stdout();
+
+    // Convert the line separator to a byte sequence.
     let sbytes = separator.as_bytes();
     let slen = sbytes.len();
+
+    // If there are more characters in the separator than in the data,
+    // we can't possibly split the data on the separator. Write the
+    // entire buffer to stdout.
+    let dlen = data.len();
+    if dlen < slen {
+        return out.write_all(&data);
+    }
+
+    // Iterate over each byte in the buffer in reverse. When we find a
+    // line separator, write the line to stdout.
+    //
+    // The `before` flag controls whether the line separator appears at
+    // the end of the line (as in "abc\ndef\n") or at the beginning of
+    // the line (as in "/abc/def").
+    let mut following_line_start = data.len();
+    for i in (0..dlen - slen + 1).rev() {
+        if &data[i..i + slen] == sbytes {
+            if before {
+                out.write_all(&data[i..following_line_start])?;
+                following_line_start = i;
+            } else {
+                out.write_all(&data[i + slen..following_line_start])?;
+                following_line_start = i + slen;
+            }
+        }
+    }
+
+    // After the loop terminates, write whatever bytes are remaining at
+    // the beginning of the buffer.
+    out.write_all(&data[0..following_line_start])?;
+    Ok(())
+}
+
+fn tac(filenames: Vec<String>, before: bool, _: bool, separator: &str) -> i32 {
+    let mut exit_code = 0;
 
     for filename in &filenames {
         let mut file = BufReader::new(if filename == "-" {
@@ -124,63 +161,8 @@ fn tac(filenames: Vec<String>, before: bool, _: bool, separator: &str) -> i32 {
             continue;
         };
 
-        // find offsets in string of all separators
-        let mut offsets = Vec::new();
-        let mut i = 0;
-        loop {
-            if i + slen > data.len() {
-                break;
-            }
-
-            if &data[i..i + slen] == sbytes {
-                offsets.push(i);
-                i += slen;
-            } else {
-                i += 1;
-            }
-        }
-        // If the file contains no line separators, then simply write
-        // the contents of the file directly to stdout.
-        if offsets.is_empty() {
-            out.write_all(&data)
-                .unwrap_or_else(|e| crash!(1, "failed to write to stdout: {}", e));
-            return exit_code;
-        }
-
-        // if there isn't a separator at the end of the file, fake it
-        if *offsets.last().unwrap() < data.len() - slen {
-            offsets.push(data.len());
-        }
-
-        let mut prev = *offsets.last().unwrap();
-        let mut start = true;
-        for off in offsets.iter().rev().skip(1) {
-            // correctly handle case of no final separator in file
-            if start && prev == data.len() {
-                show_line(&mut out, &[], &data[*off + slen..prev], before);
-                start = false;
-            } else {
-                show_line(&mut out, sbytes, &data[*off + slen..prev], before);
-            }
-            prev = *off;
-        }
-        show_line(&mut out, sbytes, &data[0..prev], before);
+        buffer_tac(&data, before, separator)
+            .unwrap_or_else(|e| crash!(1, "failed to write to stdout: {}", e));
     }
-
     exit_code
-}
-
-fn show_line(out: &mut Stdout, sep: &[u8], dat: &[u8], before: bool) {
-    if before {
-        out.write_all(sep)
-            .unwrap_or_else(|e| crash!(1, "failed to write to stdout: {}", e));
-    }
-
-    out.write_all(dat)
-        .unwrap_or_else(|e| crash!(1, "failed to write to stdout: {}", e));
-
-    if !before {
-        out.write_all(sep)
-            .unwrap_or_else(|e| crash!(1, "failed to write to stdout: {}", e));
-    }
 }