--- a/tests/by-util/test_tac.rs
+++ b/tests/by-util/test_tac.rs
@@ -1,3 +1,4 @@
+// spell-checker:ignore axxbxx bxxaxx
 use crate::common::util::*;
 
 #[test]