… tests/fixtures/tac/delimited_primes_before.expected