(Supersedes pull request #2515.)

Fixes issue #2262.

This pull request corrects the behavior of `tac -b` to match that of GNU coreutils `tac`. Specifically, this changes `tac -b` to assume *leading* line separators instead of the default *trailing* line separators.

Before this commit, the (incorrect) behavior was

    $ printf "/abc/def" | tac -b -s "/"
    def/abc/

After this commit, the behavior is

    $ printf "/abc/def" | tac -b -s "/"
    /def/abc

I have some unit tests for this option, updated some incorrect unit tests, and also added a test for the situation where the separator is more than one character long.

This implementation also slightly improves the speed of `tac`. Compare the `master` branch:
```
$ hyperfine --warmup 10 "./target/release/coreutils tac shakespeare.txt > /dev/null"
Benchmark #1: ./target/release/coreutils tac shakespeare.txt > /dev/null
  Time (mean ± σ):      48.1 ms ±   1.0 ms    [User: 30.5 ms, System: 17.5 ms]
  Range (min … max):    46.9 ms …  50.5 ms    59 runs
 ```
with this branch:
```
$ hyperfine --warmup 10 "./target/release/coreutils tac shakespeare.txt > /dev/null"
Benchmark #1: ./target/release/coreutils tac shakespeare.txt > /dev/null
  Time (mean ± σ):      42.5 ms ±   0.4 ms    [User: 25.5 ms, System: 16.9 ms]
  Range (min … max):    41.7 ms …  43.8 ms    69 runs
```
A savings of about 10% on the file https://www.gutenberg.org/files/100/100-0.txt